# Improve performance encode and decode