@peterhellberg 2017-08-16
I've now pushed `1.0.4` to https://rubygems.org/gems/hashids (with this change included)